Moving files to components folder
