+++ src/app.js
@@ -2,9 +2,9 @@
 // var rightNow = moment().format('MMMM Do YYYY, h:mm:ss a');
 // console.log(rightNow);
 
-import GameEngine from './engine.js'
-import {PLAYER_ONE, PLAYER_TWO} from './players.js'
-import Listeners from './listeners.js'
+import GameEngine from './components/engine'
+import {PLAYER_ONE, PLAYER_TWO} from './components/players'
+import Listeners from './components/listeners'
 
 GameEngine().render()
 Listeners.addListeners()
